requirements_plugin.txt
python-osc>=1.8.0
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0

